move elevation code into rastercost package
this will be generalized for elevation, tree shade etc.
